add roles

+++ api/handlers.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 from api.actions.auth import get_current_user_from_token
-from api.actions.user import _create_new_user, _delete_user, _get_user_by_id, _update_user_by_id
+from api.actions.user import _create_new_user, _delete_user, _get_user_by_id, _update_user_by_id, check_user_permissions
 from api.models import UserCreate, ShowUser, DeleteUserResponse, UpdateUserRequest, UpdateUserResponse
 from db.models import User
 from db.session import get_db
@@ -30,6 +30,13 @@ async def delete_user(
         db: AsyncSession = Depends(get_db),
         current_user: User = Depends(get_current_user_from_token)
 ) -> DeleteUserResponse:
+    user_for_deletion = await _get_user_by_id(user_id, db)
+    if user_for_deletion is None:
+        raise HTTPException(status_code=404, detail=f"User with id {user_id} not found.")
+
+    if not check_user_permissions(target_user=user_for_deletion, current_user=current_user):
+        raise HTTPException(status_code=403, detail="Forbidden.")
+
     deleted_user_id = await _delete_user(user_id, db)
     if deleted_user_id is None:
         raise HTTPException(status_code=404, detail=f"User with id {user_id} not found.")
@@ -57,13 +64,20 @@ async def update_user_by_id(
 ) -> UpdateUserResponse:
     updated_user_params = body.dict(exclude_none=True)
     if updated_user_params == {}:
-        raise HTTPException(status_code=422, detail="At least one parameter for user update should be provided")
-    user = await _get_user_by_id(user_id, db)
-    if user is None:
+        raise HTTPException(
+            status_code=422, detail="At least one parameter for user update should be provided."
+        )
+    user_for_update = await _get_user_by_id(user_id, db)
+    if user_for_update is None:
         raise HTTPException(status_code=404, detail=f"User with id {user_id} not found.")
+    if user_id != current_user.user_id:
+        if check_user_permissions(target_user=user_for_update, current_user=current_user):
+            raise HTTPException(status_code=403, detail="Forbidden.")
     try:
-        updated_user_id = await _update_user_by_id(updated_user_params=updated_user_params, session=db, user_id=user_id)
-        return UpdateUserResponse(updated_user_id=updated_user_id)
+        updated_user_id = await _update_user_by_id(
+            updated_user_params=updated_user_params, session=db, user_id=user_id
+        )
     except IntegrityError as err:
         logger.error(err)
         raise HTTPException(status_code=503, detail=f"Database error: {err}")
+    return UpdateUserResponse(updated_user_id=updated_user_id)
